requirements.txt
openai
pdfplumber
pandas
python-dotenv
ebooklib
beautifulsoup4